dags/local-requirements.txt
apache-airflow==2.6.1
apache-airflow-providers-google==10.1.1
pandas==2.0.2
pandas-gbq==0.19.2
pyarrow==9.0.0
google-cloud-storage==2.7.0
apache-airflow-providers-amazon==8.1.0
jaydebeapi==1.2.3